Custom Component Design for Android
The aim of this application is to show how can we create a component for
Android. We will create an action bar (this is not official actionbar
which is announced in Honeycomb version of Android) and will explain how
to design it in first. Then how to add it in XML and link it to code.
Finally, how to implemet listeners in order to respond of clicking items
of this actionbar.